requirements3.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
babelfish==0.5.5
beautifulsoup4==4.9.3
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorator==4.4.2
delegator.py==0.1.1
dogpile.cache==1.1.1
easydict==1.9
enzyme==0.4.1
google-api-core==1.24.1
google-api-python-client==1.7.8
google-auth==1.24.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.28.1
google-cloud-videointelligence==2.0.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.34.0
guessit==3.2.0
httplib2==0.18.1
idna==2.10
iniconfig==1.1.1
numpy==1.19.4
oauthlib==3.1.0
packaging==20.8
pbr==5.5.1
pexpect==4.8.0
pluggy==0.13.1
proto-plus==1.13.0
protobuf==3.14.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pysrt==1.1.2
pytest==6.2.1
python-dateutil==2.8.1
pytz==2020.5
rarfile==4.0
rebulk==3.0.1
requests==2.25.1
requests-oauthlib==1.2.0
rsa==4.6
six==1.15.0
soupsieve==2.1
stevedore==3.3.0
subliminal==2.1.0
toml==0.10.2
tqdm==4.56.0
uritemplate==3.0.1
urllib3==1.26.2
wincertstore==0.2
